Pick up some of the easier refactorings to remove dependencies in model code on UI code
