+++ pwiz_tools/Skyline/Model/SmallMoleculeTransitionListReader.cs
@@ -27,7 +27,6 @@ using System.Threading;
 using pwiz.Common.Chemistry;
 using pwiz.Common.SystemUtil;
 using pwiz.ProteomeDatabase.API;
-using pwiz.Skyline.EditUI;
 using pwiz.Skyline.Properties;
 using pwiz.Skyline.Util;
 using pwiz.Skyline.Util.Extensions;
@@ -2129,5 +2128,13 @@ namespace pwiz.Skyline.Model
             Thread.CurrentThread.CurrentUICulture = currentUICulture;
             KnownHeaderSynonyms = knownColumnHeadersAllCultures;
         }
+    }
+
+    public class PasteError
+    {
+        public String Message { get; set; }
+        public int Line { get; set; }
+        public int Column { get; set; }
+        public int Length { get; set; }
     }
 }
