+++ pwiz_tools/Skyline/SettingsUI/PeptideSettingsUI.cs
@@ -632,7 +632,7 @@ namespace pwiz.Skyline.SettingsUI
                 list.SetValue(calcNew);
                 // Automatically add new RT regression using this calculator
                 var regressionName = Helpers.GetUniqueName(calcNew.Name, name => !_driverRT.List.Contains(r => Equals(r.Name, name)));
-                var regression = new RetentionTimeRegression(regressionName, calcNew, null, null, EditRTDlg.DEFAULT_RT_WINDOW, new List<MeasuredRetentionTime>());
+                var regression = new RetentionTimeRegression(regressionName, calcNew, null, null, ImportPeptideSearch.DEFAULT_RT_WINDOW, new List<MeasuredRetentionTime>());
                 Settings.Default.RetentionTimeList.Add(regression);
                 _driverRT.LoadList(regression.GetKey());
             }
